@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { Component, OnInit } from '@angular/core';
+import { Component, inject, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { User } from '../user';
 import { UserService } from '../user.service';
@@ -12,8 +12,8 @@ import { UserService } from '../user.service';
 })
 export class UserListComponent implements OnInit {
   users: User[] = [];
-
-  constructor(private router: Router, private userService: UserService) {}
+  private router = inject(Router);
+  private userService = inject(UserService);
 
   ngOnInit(): void {
     this.getUsers();
